+++ packages/shared/common/src/datasource/index.ts
@@ -1,14 +1,21 @@
+import { Backoff, DefaultBackoff } from './Backoff';
+import { CompositeDataSource } from './CompositeDataSource';
 import { DataSourceErrorKind } from './DataSourceErrorKinds';
 import {
   LDFileDataSourceError,
+  LDFlagDeliveryFallbackError,
   LDPollingError,
   LDStreamingError,
   StreamingErrorHandler,
 } from './errors';
 
 export {
+  Backoff,
+  CompositeDataSource,
+  DefaultBackoff,
   DataSourceErrorKind,
   LDFileDataSourceError,
+  LDFlagDeliveryFallbackError,
   LDPollingError,
   LDStreamingError,
   StreamingErrorHandler,