+++ packages/shared/sdk-server/src/store/InMemoryFeatureStore.ts
@@ -1,32 +1,23 @@
+import { internal } from '@launchdarkly/js-sdk-common';
+
 import { DataKind } from '../api/interfaces';
 import {
-  LDFeatureStore,
   LDFeatureStoreDataStorage,
   LDFeatureStoreItem,
   LDFeatureStoreKindData,
   LDKeyedFeatureStoreItem,
+  LDTransactionalFeatureStore,
 } from '../api/subsystems';
 
-export default class InMemoryFeatureStore implements LDFeatureStore {
+export default class InMemoryFeatureStore implements LDTransactionalFeatureStore {
   private _allData: LDFeatureStoreDataStorage = {};
+
+  // this tracks the last received selector, which may not be present
+  private _selector: string | undefined;
 
   private _initCalled = false;
 
-  private _addItem(kind: DataKind, key: string, item: LDFeatureStoreItem) {
-    let items = this._allData[kind.namespace];
-    if (!items) {
-      items = {};
-      this._allData[kind.namespace] = items;
-    }
-    if (Object.hasOwnProperty.call(items, key)) {
-      const old = items[key];
-      if (!old || old.version < item.version) {
-        items[key] = item;
-      }
-    } else {
-      items[key] = item;
-    }
-  }
+  private _initMetadata?: internal.InitMetadata;
 
   get(kind: DataKind, key: string, callback: (res: LDFeatureStoreItem | null) => void): void {
     const items = this._allData[kind.namespace];
@@ -52,20 +43,82 @@ export default class InMemoryFeatureStore implements LDFeatureStore {
     callback?.(result);
   }
 
-  init(allData: LDFeatureStoreDataStorage, callback: () => void): void {
-    this._initCalled = true;
-    this._allData = allData as LDFeatureStoreDataStorage;
-    callback?.();
+  init(
+    allData: LDFeatureStoreDataStorage,
+    callback: () => void,
+    initMetadata?: internal.InitMetadata,
+  ): void {
+    this.applyChanges(true, allData, callback, initMetadata);
   }
 
   delete(kind: DataKind, key: string, version: number, callback: () => void): void {
-    const deletedItem = { version, deleted: true };
-    this._addItem(kind, key, deletedItem);
-    callback?.();
+    const item: LDKeyedFeatureStoreItem = { key, version, deleted: true };
+    this.applyChanges(
+      false,
+      {
+        [kind.namespace]: {
+          [key]: item,
+        },
+      },
+      callback,
+    );
   }
 
   upsert(kind: DataKind, data: LDKeyedFeatureStoreItem, callback: () => void): void {
-    this._addItem(kind, data.key, data);
+    this.applyChanges(
+      false,
+      {
+        [kind.namespace]: {
+          [data.key]: data,
+        },
+      },
+      callback,
+    );
+  }
+
+  applyChanges(
+    basis: boolean,
+    data: LDFeatureStoreDataStorage,
+    callback: () => void,
+    initMetadata?: internal.InitMetadata,
+    selector?: string,
+  ): void {
+    if (basis) {
+      this._initCalled = true;
+      this._allData = data;
+      this._initMetadata = initMetadata;
+    } else {
+      const tempData: LDFeatureStoreDataStorage = {};
+      // shallow copy to protect against concurrent read
+      Object.entries(this._allData).forEach(([namespace, items]) => {
+        tempData[namespace] = { ...items };
+      });
+
+      Object.entries(data).forEach(([namespace, items]) => {
+        Object.keys(items || {}).forEach((key) => {
+          let existingItems = tempData[namespace];
+          if (!existingItems) {
+            existingItems = {};
+            tempData[namespace] = existingItems;
+          }
+          const item = items[key];
+          if (Object.hasOwnProperty.call(existingItems, key)) {
+            const old = existingItems[key];
+            // TODO: SDK-1046 - Determine if version check should be removed
+            if (!old || old.version < item.version) {
+              existingItems[key] = { key, ...item };
+            }
+          } else {
+            existingItems[key] = { key, ...item };
+          }
+        });
+      });
+
+      this._allData = tempData;
+    }
+
+    this._selector = selector;
+
     callback?.();
   }
 
@@ -80,5 +133,13 @@ export default class InMemoryFeatureStore implements LDFeatureStore {
 
   getDescription(): string {
     return 'memory';
+  }
+
+  getInitMetaData(): internal.InitMetadata | undefined {
+    return this._initMetadata;
+  }
+
+  getSelector(): string | undefined {
+    return this._selector;
   }
 }